Create Payload class that handles attributions
* The Attributions module is now redundant, though I should now see what
happens when we try to put in an extra attribute to payload.

We get this error:
===========================================================================
 - ActiveModel::UnknownAttributeError - unknown attribute 'flavour' for View.:
===========================================================================

* Reset the extract method on Payload

We currently need to use the attributions module to set the attributes,
we need to do this as otheriwse we get an error that the attribute does
not exist.

* We also remove the attributions module from the webapp itself,
as it can already be used from within the payload class.